requirements.txt
pandas==2.1.0
jpholiday-0.1.10
requests==2.31.0